Minor Atira Improvements & Fixes (#5166)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
- Moves some lightswitches to be in more accessible places
- Adds a windoor to the cargo bay/waiting room
- Removes some uneccesary crate shelvers
- Moves some objects slightly that were blocking APCs visually
- Fills in some empty tables in the bridge and cryo room
- Fixes faxes being PGF fax machines , for some reason
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Lightswitches aren't in weird places anymore, harder to ruin the atmos
in the cargo bay, ship has radios now other than the ones you spawn with
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: minor Atira-class QoL tweaks
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->