Animate elzuose tails (#5189)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

animates the elzu tails to be more than their current 2 frame one. full
wag seemed to be more popular of the two initial animations


https://github.com/user-attachments/assets/fc9a690f-b001-4a83-926c-46faea435e3b



## Why It's Good For The Game

current tail wag for the longer tails is i assume more of a placeholder,
gives the tail wags more of a standard animation to the rest of the
tails