Check Fixing
Capitalizes all the rooms that I just didn't capitalize for some reason. \\ Removes the subshuttle dock for the moment because I'm not 100% on what specifically went wrong since it looks identical in the .dmm's text as ship shuttle docks and should fit without landing on any walls. \\  Removes the double shutters and dirt hiding in the wall.