First attempt at addressing criticisms
Removes Budgetcuts.dmm, its area, and its entry in rockplanet.dm.

Rearranges the Frontiersman Trench Complex and extends it in the X direction to accommodate a true outdoors landing pad and the surrounding trench. Adds few extrusions with turrets and small mud-trench lines to get to them.

The interior should now be atmos-safe, with faux-airlocks similar to tesla lab. In my head they just slept with their oxy cans on because they're buffoons like that but it's probably better to do proper atmos. The LZ, now a warehouse, has holofields because of it.

Makes a new Frontiersman Boss subtype, the Frontiersman Captain, who has a wasp. +1 on the burst but extra spread compared to the trooper version. Also adds internals and neutered versions of the Frontiersman Commander, the rifle wielding boss, because I thought it would be nice for consistency.

To keep Ross in line with Commander Tiyak as a 'custom' boss I created the code\modules\ruins\rockplanet_ruin_code\somme.dm and stuck him in there with the 100% drop chance. Randomname gen on death is fine since Ross is probably a pseudonym anyway.