Fixes Areas and encircles the LZ
Forgot to make a change to the Trenchline area to accompany a change in the shape of the outbound trenches and circled the entire LZ platform so even the most foolish Frontiersman can recognize that the 'stand clear' goes all the way. Also changes the wood type in a few sections to provide better room-to-room distinction.