Second Round of Criticism
Added a non-holofield entry to the warehouse that's locked to begin with by a shutter over the door. The override button for it is in Ross' quarters same as the cargo shutters used to be.

The Inteq corpses are now a bit more organized, insofar as not just being lined up against the wall, and are in the process of being stripped for equipment. The sneaksuit is because I think it's cute and undifferentiated except by sprite but I can put it to a normal Inteq uniform if it's TOO cute.

I considered making a 'trench' subtype in somme.dm for the retreat variable change since other people might want to use it as well, but decided that would clog up the prefab list too much and just varedited them.

A couple of the turrets (E, NE, SW) were changed from spitter to pounder. E and NE to give an avenue of 'less resistance' for return visitors and SW so it can act as a 'hey this ruin exists and has danger' marker for crews.

Floors in the warehouse section were changed to dark tile and the section with the fuel tank was changed to be an exo recharger with a broken power-loader since the warehouse would ICly be getting and sending regular raid and supply shipments.

Let's see if I put dirt into any walls again.