Map changes + Frontiersman Sentries
Changes sections of the Trench Complex according to feedback. Adding several entrances to the trenchline and removal of shanks. Also puts the air pump+manifold into the correct order. The Armoury now starts open, and Ross has windows.

All the Frontiersmen who should be roaming the trenches should now have their structure_smash set to 0, so instead of hitting and getting stuck on the platforms they seem, emphasis on the word, to simply pass through them now.

A commissioner is replaced by a Mongrel. There will be no ammo for it since the Frontiersmen don't have a 10mm gun that they would reasonably stock.

Creates a 'new' Frontiersman type, the Frontiersman Sentry. It's just a Frontiersman Sharpshot but they have a longer engagement range to reflect a 'proper sniper' and won't move at all. Essentially a fleshy turret if you think about it a little. This was mostly done because for some reason vision_range would reset from 14 to the generic type's 9 whenever the snipers picked and killed/lost a target. If I knew better how this worked I'd probably change it at the ground level but setting it to their own bespoke type worked in my testing. (Technically they only begin to engage at 13 tiles, not 14. But if you step into engagement and back out to 14 tiles they'll still shoot. I assume it's a discrepancy in how the distances are calculated).