Fixes ICW Base I guess?
Apparently replacing the fax on ICW base deleted the TComm room's walls? I don't know how this happened? It wasn't even all no diagonal walls it was JUST the no diagonal walls that weren't otherwise otherwise occupied by things like turrets and what have you. I'm befuddled and confused. I don't know how adding the fax caused this but it must have because it's in the commit as well and I didn't think too much of it at the time.