Reverts #1885, Depends #5236 (#5237)
Basically requires
[#5236.](https://github.com/shiptest-ss13/Shiptest/pull/5236)
**Byond ver => 515.1644 on server is required.**

Closes [#3692](https://github.com/shiptest-ss13/Shiptest/issues/3692)

In summary, C style loops cause the DM virtual machine to have to
perform context switches and extra ops on each iteration. In style loops
do not. There is a 5 to 10x difference in speed each time the loop
starts. C style loops had to be used because of a bug that got fixed in
1644, making the workaround no longer necessary.

This isn't going to be a revolutionary speed increase for ships landing
and taking off as the contents of these loops are heavier than the time
it takes to restart the loop, but a little speedup might be seen as
there are a lot of iterations occurring.

Tested this and #5236 on 515.1647 and 516.1667 by taking off and
landing.

:cl:
refactor: Byond version 515.1644 fixed a nasty bug with in style loops
that led to SGTs. Reverting workaround code as prior PR requires
515.1644 or higher.
/:cl: