Gas mask resprite+sec hailer resprite/reflavoring  (#5314)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Sec hailer is reflavored to a generic half mask, also adds it to loadout
so its not literally nowhere.
Species variants include kepi, sarathi and human for the half mask and
kepi and human for the full mask (I had 0 ideas for the sarathi mask)
<img width="1485" height="694" alt="image"
src="https://github.com/user-attachments/assets/86947cd3-0717-43ac-bffc-40287e19b70c"
/>


## Why It's Good For The Game
The old sprites were pretty ass. These should be nicer, and as for the
half mask in loadout, its literally just a reflavored breath mask. Gas
masks as they are now are only valuable with eye protection, which this
has none.

## Changelog

:cl:
add: Gas mask resprite, sec hailer resprite and reflavoring
add: half mask to loadout
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->