Solarbundswaffenkammer makes their guns useful again (#5239)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
most of it is in the changelog
I increased the solarian guncells to 15000 charge, and tweaked namely
the GAR as said in the changelog to have a reasonable tradeoff of
cell-to-shot ratio of 96 rounds standard, and 64 HC. I didn't tweak the
other guns, because their energy consumption is fine as it is, with the
increased full-charge state.
This isn't like. Gamebreaking because nobody plays Sol, but makes them
actually able to shoot their own guns instead of looking like 2-bit
gangsters with SKM's that actually work.

## Why It's Good For The Game
I get nobody plays SolCon but when you can't even shoot a full magazine
out of your GAR I think it's a little silly.
Also you can buy everything properly and not get ferromagnetic HC lances
ONLY
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Ferromagnetic Lances BACK into cargo.
balance: Solarian guns now have better rates of cell decrease, namely
the GAR, which can fire 3 standard magazines, or 2 HC magazines.
balance: Ferromagnetic rounds now stack better and reasonably. Slugs:
10; Lance: 16; Pellets: 22, now you can load your guns actually with a
reasonable amount of boxes!
fix: Solarian HC box sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->