Assorted Hairs: The Hairening (#5411)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR resprites a whopping 21 haircuts! Wow! That's a lot of hair!
Styles include:
- Nearly every homestuck hair except for "Eight" (sorry it was quite
literally a very subtle edit of FLBH, which is why I deleted it)
- Bowlcut (the "2" variant has been annihilated)
- Big Afro
- Afro
- FLBH
- Long Bedhead
- Dandy Pompadour (the "basic" variant has been mulched)
- Ronin
- Mohawk
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
New hairs are great! Huge credits for @spookytwirl for bringing this to
the finish line!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy, Spookywastaken/Spookytwirl
del: Certain duplicate hair styles have been erased.
imageadd: A variety of hairs have been resprited!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: FalloutFalcon <falloutfalconplays@gmail.com>