Refactors IPC chassis code to be less evil (#5409)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

IPC chassis code has been reworked to be far easier to add more variants
in the future that work differently from any that already exist, and to
kill the need to add redundant sprites for the entire body just to add a
variant with a slightly different head sprite.

This also adds the ability to fully customize each individual part of an
IPC's chassis, and select any IPC chassis style as a prosthetic limb
option.

Additionally, bodyparts are now stored as an associated list to prevent
layering issues that became much more apparent with the above changes,
and to allow for species with more or fewer limbs than normal to be
potentially added in the future. Non-digitigrade sarathi had to be
sacrificed for this functionality, but probably won't be missed.

## Why It's Good For The Game

Kills bad code and redundant sprites with hammers, also more character
customization is pretty neat.

## Changelog

:cl:
add: IPC frames and prosthetic limbs can now be fully customized
fix: Fixed bodyparts layering incorrectly
del: Removed non-digitigrade sarathi for the sake of less evil code
refactor: Refactored bodyparts code to allow for species with more or
fewer limbs than normal
refactor: Refactored IPC chassis code
imagedel: Removed redundant IPC chassis sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->