removes stuff that shouldn't be there (#5470)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
noticed these salvaging and they annoyed me
the shutters in the n+s kitchen in the mining facility were doubled up
there was a wall inside of the bridge windows in the crashed starwalker
## Why It's Good For The Game
immersion ruining mapping mistake that must be purged
## Changelog

:cl: Goat
fix: removed random wall that was inside of a window on crashed
starwalker
fix: the shutters are no longer doubled up in the n+s rock planet mining
facility kitchen
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->