Changes the orientation of the briefing room on the Kali Andhi (#5437)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This map change makes it so that the briefing room has its chairs and
podium moved so that the sidestates of characters in the room are shown
instead of the front state and the backstate.

Before:
<img width="717" height="575" alt="image"
src="https://github.com/user-attachments/assets/64f1fa96-1243-4a4e-b38f-554aba879691"
/>

After:
<img width="570" height="575" alt="image"
src="https://github.com/user-attachments/assets/2620c3b5-923c-4b7b-9559-308aba9636f9"
/>

<img width="517" height="511" alt="image"
src="https://github.com/user-attachments/assets/b160e02e-4a80-490d-bf94-8d53363824b5"
/>

## Why It's Good For The Game

It can be a little awkward to have your backsprite be displayed when
you're in a briefing, or giving one. Changing the orientation of the
briefing room enables everyone to show their sidestate, which tends to
be better.

## Changelog

:cl:
refactor: modified the orientation of the Kali Andhi's briefing room.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zimon9 <122945887+zimon9@users.noreply.github.com>