Adds hardsuit jetpack upgrades to cargo and makes them a little more versatile (#5497)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds hardsuit jetpack upgrades to the cargo market for 2000
credits, and it changes their behavior so that they are no longer
restricted to just taking air from a tank in the suit storage slot. They
work on a priority list now, first drawing air from the suit storage
slot, then checking the back slot, then the belt, then the left pocket,
and finally the right pocket, for a tank to draw propellant from.

## Why It's Good For The Game

This is a minor little addition that provides a more accessible option
if one doesn't want to make use of a bulky jetpack.

## Changelog

:cl:
add: Added hardsuit jetpack upgrades to cargo for 2000 credits each
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->