Fixes every right click action in the game being broken + adds a few more (#5584)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds right-click actions for:
- Suit sensors
- Using modular computers
- Unlocking APCs and air alarms
- Deconstructing tables, windows, and walls
- Welding airlocks closed
- Toggling gun safeties
- Opening/closing fire extinguisher cabinets
- Wrapping items in tape
- Renaming paystands
- Medical malpractice
- Tipping medibots over

Fixes right-click actions for:
- Every single one that was added. I am impressed with the lack of
testing.

If you think of something that should be done with right click, make a
comment about it and I'll add it.

## Why It's Good For The Game

Fix + QoL

## Changelog

:cl:
add: Added more right-click actions
fix: Fixed existing right-click actions not working
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->