Fixes gender being reset when loading a character with none selected (#5589)
## About The Pull Request

Fixes a bug that causes a character's gender to be reset to male when
loading it, if that character had none selected.

## Why It's Good For The Game

Gender

## Changelog

:cl:
fix: Fixed gender being set to male when loading a character with none
selected
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->