Standardizes premium ion engines (#5519)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Changes premium ion engines so that they are the same as standard ion
engines, only differing by starting with tier 3 parts upon mapload.

## Why It's Good For The Game

This is a continuation of an earlier PR that changes the behavior of ion
engines, and it opens the door for custom ion engines of different make.

## Changelog

:cl:
balance: standardized premium ion engines with standard ion engines
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->