Adds inacusiate, oculine, and charcoal to the outpost medvender (#5574)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds autoinjectors for oculine and inacusiate to the outpost
medvender, as well as a pill bottle filled with three pills of charcoal
(the charcoal/less pill bottle subtype)

<img width="421" height="383" alt="image"
src="https://github.com/user-attachments/assets/b5f97f7d-3636-40b8-b28d-a3482e7d00b8"
/>

## Why It's Good For The Game

There currently does not exist an accessible way to obtain oculine and
inacusiate, and this PR aims to address this issue. There's also nothing
in the medical vender that directly treats toxins at all, so a pill
bottle for charcoal was added.

## Changelog

:cl:
add: Adds a charcoal pill bottle, and oculine/inacusiate autoinjectors
to the outpost medical vender
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->