516 compile fix (#5567)
## About The Pull Request

Removes the use of integers as associated list keys to fix a number of
compiler errors on 516.

## Why It's Good For The Game

Yes

## Changelog

:cl:
code: Fixed compile error on 516
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->