M20 price increase and pinscher back in cargo (#5697)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

M20 price increased from 1250 to 1500, pinscher returned to cargo so you
can buy it

## Why It's Good For The Game

Price increase was suggested and .44 is a good caliber
Pinscher is a reskin of an indie-available gun so it not being in cargo
is weirfd and the change was undocumented

## Changelog

:cl:
balance: m20 auto elite price increased from 1250 to 1500
add: the pinscher has been returned to cargo
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Theos <theubernyan@gmail.com>