Gives mounted widebands a dedicated use key, and prevents intercom use from sending messages through a wideband as well (#5573)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This pr adds a new use key for widebands, this being :w -- a few
modifications were also made so that using the :i key when next to both
a wideband and a intercom no longer sends a message through both of them

<img width="586" height="360" alt="image"
src="https://github.com/user-attachments/assets/c9baec3d-6d1f-45f1-95e5-54df45109305"
/>

## Why It's Good For The Game

Previously, when you wanted to use _just_ a wideband or _just_ an
intercom when multiple were near each other, you'd need to activate the
microphone for each one manually. This had its own problems, as you had
to ensure that no one else was speaking when the microphone was on, and
it was a hassle to do, requiring three separate actions to send one
message. This should make it so that widebands are now just as easy to
use as an intercom.

## Changelog

:cl:
add: Added a dedicated wideband intercom key
balance: prevents the intercom key from sending messages through
widebands
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->