Paraplegia improvements (#5674)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR allows the use of UIs from things like radios and consoles while
floored, and it also allows upright movement when in zero-gravity for
those without the ability to move their legs

Piggybacking was also modified so that if someone that wanted to
piggyback someone else was buckled into something like a wheelchair,
they don't have to unbuckle and lie on the floor in order to get on the
other person's back -- they can just climb on directly now

Thank you to Erika for help with the signals

## Why It's Good For The Game

It makes sense to be able to use a radio or a tablet while lying down, I
think, and the allowance of upright movement in zero-G allows characters
without the use of their legs to simulate spacewalking, where gravity no
longer holds them down

## Changelog

:cl: ErikaFox
add: upright movement when floating or flying and paraplegic
del: removed the restriction on the floored trait that didn't allow for
the use of UIs while on the floor
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->