Replaces the use of ambrosia in salads with cabbage (#5631)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Replaces Ambrosia Vulgaris and Ambrosia Deus in the crafting recipe of
salads with cabbage, and replaces Ambrosia in fridges with cabbage as
well.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Last time I was on a Sunskipper and served a salad, it had drugs in it.
I think drugging your customers is bad. Ambrosia is eluded to being a
drug, so using it for most salads is probably bad. I would've used
lettuce, but apparently we don't even have lettuce. No need to have
drugs in your fridge if there's nothing to make with them.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Salad recipies now use cabbage instead of ambrosia. Fridges
appropriately modified.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->