Okay I think this is a better version of my previous pr? (still fixing formal/epipens) (#5681)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Makes it so that formaldahide (thats not how you spell that is it?) will
not do histamine unless you have over 5 units of it in your system
## Why It's Good For The Game
This prevents epipens from killing you
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:

balance:formaldehyde can now only trigger its histamine producing effect
if there is more than 5u of formaldehyde in your body (epipens contain
3u)

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->